Added types for states, events, updated extension and imports.
