+++ Frontend/src/pages/Login.tsx
@@ -1,0 +1,114 @@
+import { useState, useContext, ChangeEvent, FormEvent } from "react";
+import { Link } from "react-router-dom";
+
+import userContext from "@src/context/Users/userContext";
+
+interface UserLogin {
+  email: string;
+  password: string;
+}
+
+const Login = () => {
+  /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+  /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+
+  // VARIABLES
+
+  const context = useContext(userContext);
+  const { logIn } = context;
+
+  const [userData, setUserData] = useState<UserLogin>({
+    email: "",
+    password: "",
+  });
+
+  /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+  /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+
+  // METHODS
+
+  const onchange = (e: ChangeEvent<HTMLInputElement>) => {
+    e.preventDefault();
+    setUserData({ ...userData, [e.target.name]: e.target.value });
+  };
+
+  const handleLogIn = async (e: FormEvent<HTMLButtonElement>) => {
+    e.preventDefault();
+    logIn(userData);
+    setUserData({
+      email: "",
+      password: "",
+    });
+  };
+
+  /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+  /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+
+  // RETURN
+
+  return (
+    <div className="max-w-screen-lg mx-auto">
+      <div className="pb-5 -mt-5 text-center">
+        <h1 className="text-3xl italic font-bold">DevNotes</h1>
+        <p className="text-base font-semibold">Your Notes secured on the cloud</p>
+      </div>
+      <div className="p-2 px-4 border border-light rounded-4 custom-shadow mx-[7vw]">
+        <h2 className="py-2 text-lg font-semibold text-center">Log In to DevNotes</h2>
+        <form>
+          <div className="mb-3 form-floating ">
+            <input
+              type="email"
+              className=" form-control rounded-4 bg-opacity-10"
+              id="email"
+              name="email"
+              value={userData.email}
+              onChange={onchange}
+              placeholder="E-mail"
+            />
+            <label htmlFor="floatingInput">Email address</label>
+          </div>
+          <div className="mb-3 form-floating">
+            <input
+              type="password"
+              className="form-control rounded-4 bg-opacity-10"
+              id="password"
+              name="password"
+              value={userData.password}
+              onChange={onchange}
+              placeholder="Password"
+            />
+            <label htmlFor="floatingPassword">Password</label>
+          </div>
+
+          <div className="flex py-2 justify-content-between">
+            <div>
+              <button
+                type="submit"
+                className="btn-sm btn-bg-custom "
+                style={{ maxWidth: "8rem" }}
+                onClick={handleLogIn}
+              >
+                Log In
+              </button>
+            </div>
+            <div className="py-2">
+              <p>
+                Don't have an Account?
+                <Link type="submit" className="mx-1 custom-link" to="/signup">
+                  <b> Sign Up Here</b>
+                </Link>
+              </p>
+            </div>
+          </div>
+        </form>
+      </div>
+    </div>
+  );
+};
+
+/* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+/* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
+
+// EXPORT
+
+export default Login;
