+++ Frontend/src/pages/SIgnup.tsx
@@ -1,19 +1,23 @@
-import { useState, useContext } from "react";
+import { useState, useContext, ChangeEvent, FormEvent } from "react";
 import { Link } from "react-router-dom";
-import userContext from "../context/Users/userContext";
 
-const Login = () => {
+import userContext from "@src/context/Users/userContext";
+import { NewUser } from "@srctypes";
+
+const SIgnup = () => {
   /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
   /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
 
   // VARIABLES
 
   const context = useContext(userContext);
-  const { logIn } = context;
+  const { signUp } = context;
 
-  const [userData, setUserData] = useState({
+  const [userData, setUserData] = useState<NewUser>({
+    name: "",
     email: "",
     password: "",
+    cpassword: "",
   });
 
   /* ----------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------------- */
@@ -21,17 +25,19 @@ const Login = () => {
 
   // METHODS
 
-  const onchange = (e) => {
+  const onchange = (e: ChangeEvent<HTMLInputElement>) => {
     e.preventDefault();
     setUserData({ ...userData, [e.target.name]: e.target.value });
   };
 
-  const handleLogIn = async (e) => {
+  const handleSignUp = async (e: FormEvent<HTMLButtonElement>) => {
     e.preventDefault();
-    logIn(userData);
+    signUp(userData);
     setUserData({
+      name: "",
       email: "",
       password: "",
+      cpassword: "",
     });
   };
 
@@ -41,30 +47,43 @@ const Login = () => {
   // RETURN
 
   return (
-    <div className="w-full">
-      <div
-        style={{
-          paddingTop: "7vh",
-        }}
-      ></div>
+    <div className="custom-text">
+      <div style={{
+        paddingTop: "7vh"
+      }}>
 
-      <div className="pb-2 maxHeight-30rem">
+      </div>
+
+      <div className="pb-2">
         <div className="container text-center lh-lg ">
           <h1 className="container fw-bold">
             <i> Ur Notes </i>
           </h1>
           <p className="fs-6">Your Notes secured on the cloud</p>
         </div>
-        <div
-          className="p-2 px-4 border border-light rounded-4 custom-shadow "
-          style={{
-            marginTop: "2vh",
-            marginLeft: "7vw",
-            marginRight: "7vw",
-          }}
-        >
-          <h2 className="container py-2 mb-3 text-center">Log In to DevNotes</h2>
+        <div className="p-2 px-4 border border-light rounded-4 custom-shadow" style={{
+          marginTop: "2vh",
+          marginLeft: "7vw",
+          marginRight: "7vw",
+        }}>
+
+
+          <h2 className="container py-2 mb-3 text-center">
+            Sign Up to DevNotes
+          </h2>
           <form>
+            <div className="mb-2 form-floating ">
+              <input
+                type="text"
+                className="form-control rounded-4 bg-opacity-10"
+                id="name"
+                name="name"
+                value={userData.name}
+                onChange={onchange}
+                placeholder="name"
+              />
+              <label htmlFor="floatingInput">Full Name</label>
+            </div>
             <div className="mb-3 form-floating ">
               <input
                 type="email"
@@ -90,22 +109,35 @@ const Login = () => {
               <label htmlFor="floatingPassword">Password</label>
             </div>
 
-            <div className="flex py-2 justify-content-between">
+            <div className="mb-3 form-floating">
+              <input
+                type="password"
+                className="form-control rounded-4 bg-opacity-10"
+                id="cpassword"
+                name="cpassword"
+                value={userData.cpassword}
+                onChange={onchange}
+                placeholder="Confirm Password"
+              />
+              <label htmlFor="floatingPassword">Confirm Password</label>
+            </div>
+
+            <div className="flex justify-content-between">
               <div>
                 <button
                   type="submit"
                   className="btn-sm btn-bg-custom "
                   style={{ maxWidth: "8rem" }}
-                  onClick={handleLogIn}
+                  onClick={handleSignUp}
                 >
-                  Log In
+                  Sign Up
                 </button>
               </div>
               <div className="py-2">
                 <p>
-                  Don't have an Account?
-                  <Link type="submit" className="mx-1 custom-link" to="/signup">
-                    <b> Sign Up Here</b>
+                  Have an Account?
+                  <Link type="submit" className="mx-1 custom-link" to="/login">
+                    <b> Log In Here</b>
                   </Link>
                 </p>
               </div>
@@ -113,7 +145,7 @@ const Login = () => {
           </form>
         </div>
       </div>
-      {/* <div className="py-3"> </div> */}
+      <div className="py-3"> </div>
     </div>
   );
 };
@@ -123,4 +155,4 @@ const Login = () => {
 
 // EXPORT
 
-export default Login;
+export default SIgnup;
